requirements.txt
django
django-crispy-forms
django-debug-toolbar
django-extensions
django-filter
djangorestframework
drf-yasg
djangorestframework-xml
djangorestframework-csv
djangorestframework-simplejwt

celery

requests
beautifulsoup4
lxml

ipython

# flake8
flake8
flake8-blind-except==0.2.0
flake8-breakpoint==1.1.0
flake8-bugbear==21.9.2
flake8-builtins==1.5.3
flake8-debugger==4.0.0
flake8-eradicate==1.2.0
flake8-mutable==1.2.0
flake8-print==4.0.0
flake8-pytest-style==1.5.1
flake8-use-fstring==1.3
flake8-variables-names==0.0.4
flake8-import-order
pep8-naming==0.12.1

python-dotenv==0.20.0